Get rid of "shortId" concept
Short ID was used previously as a replacement for calendar ID to
save on space.

Short ID was bug prone, as it used just an index of the calendar
in the API response, but the order of calendars changes.

Sorting by ID before assigning Short ID would be buggy too, as creating
a new calendar (which would receive a random ID), would offset by 1
all the calendars after the current one.

Can not use date of calendar creation either, as user may be added
to a calendar that was created earlier than one of the user's personal
calendars.

In addition to the Short ID bugs, when plotting the data, the current
data structure was posing some difficulties.

Thus, I changed the data structure to make it more efficient, remove
the need for Short Id and make plotting data easier