Fix test depending on daylights savings
Fix #114

Fixed by removing the test for time zone offset because this is actually
a non issue.

The test depends on daylights savings because it runs under users local
time, rather than mocks the time

Mocking the time would fix this issue.

In places were consistent results are needed, the time is already
mocked.

Everywhere else, it is best not to mock time in order to better
approximate real world conditions (e.g, would help detect the cases when
some test fails on a given day of the week on in a given part of day)