Fix extension not working after first authentication
Until page reload

Happens because when the page loads, the extension needs to identify
which calendars are visible. Until calendars are fetched, it can't do
that. Calendars are't fetched until you are authenticated.

However, once calendars are fetched, the useVisibilityChangeSpy
does not trigger it's useEffect again, thus it keeps on waiting for
calendars to be fetched.

Fixes #211