fix(DomReading): make unknown calendar ID error more descriptive
It seems that when I toggled visibility of a calendar, the events from
the newly disabled calendar were still temporary visible, causing an
error.

If that is the case, then that should already be handled by the fact
that I read DOM up to 10 times on failure to account for
inconsistencies.

However, just in case that isn't the cause, let's make the error message
more descriptive.