refactor(BetterEditRecurring): simplify logic
There is a bug here that sometimes causes duplicate
"Edit repeated event" dialog to occur - I tried simplifying the logic
and debugging, but couldn't figure out a solution.

The best I figure out is that the .click() event (which is correctly
called on the previous dialog) is being ignored, and so instead of
previous dialog being closed, its item becomes unselected, and a new
dialog is opened.