refactor(DomReading): reduce dependence on DOM
I don't need to read the times from the DOM at all as that can be
retrieved from the aria string.

Would have been even better if I didn't need to read the summary, but
unfortunately there is no way around that (without knowing the summary,
it is ambiguous whether aria string is for a two-day event or whether
the summary simply includes numbers).